+++ demos/world/src/main/java/is/codion/framework/demos/world/ui/ContinentTablePanel.java
@@ -20,8 +20,6 @@ final class ContinentTablePanel extends EntityTablePanel {
 
   @Override
   protected Controls createPopupMenuControls(List<Controls> additionalPopupMenuControls) {
-    return control(ControlCode.REFRESH)
-            .map(Controls::controls)
-            .orElse(null);
+    return Controls.controls(control(ControlCode.REFRESH));
   }
 }