+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTablePanel.java
@@ -266,7 +266,6 @@ public class EntityTablePanel extends JPanel {
   }
 
   private static final int FONT_SIZE_TO_ROW_HEIGHT = 4;
-  private static final Control NULL_CONTROL = Control.control(() -> {});
   private static final Function<SwingEntityTableModel, String> DEFAULT_STATUS_MESSAGE = new DefaultStatusMessage();
 
   private final State conditionPanelVisibleState = State.state(CONDITION_PANEL_VISIBLE.get());
@@ -609,12 +608,24 @@ public class EntityTablePanel extends JPanel {
 
   /**
    * @param controlCode the control code
-   * @return the control associated with {@code controlCode} or an empty Optional if no control is available
+   * @return true if this table panel contains a control assocated with the given {@code controlCode}
    */
-  public final Optional<Control> control(ControlCode controlCode) {
-    Control control = controls.get(requireNonNull(controlCode));
+  public final boolean containsControl(ControlCode controlCode) {
+    return controls.get(requireNonNull(controlCode)) != null;
+  }
 
-    return control == NULL_CONTROL ? Optional.empty() : Optional.ofNullable(control);
+  /**
+   * @param controlCode the control code
+   * @return the control associated with {@code controlCode}
+   * @throws IllegalArgumentException in case no control is associated with the given {@code controlCode}
+   * @see #containsControl(EntityTablePanel.ControlCode)
+   */
+  public final Control control(ControlCode controlCode) {
+    if (!containsControl(controlCode)) {
+      throw new IllegalArgumentException(controlCode + " control not available in panel: " + this);
+    }
+
+    return controls.get(controlCode);
   }
 
   /**
@@ -746,37 +757,37 @@ public class EntityTablePanel extends JPanel {
    * Sets up the default keyboard actions.
    */
   protected void setupKeyboardActions() {
-    control(ControlCode.REQUEST_TABLE_FOCUS).ifPresent(control ->
+    getControl(ControlCode.REQUEST_TABLE_FOCUS).ifPresent(control ->
             KeyEvents.builder(VK_T)
                     .modifiers(CTRL_DOWN_MASK)
                     .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                     .action(control)
                     .enable(this));
-    control(ControlCode.SELECT_CONDITION_PANEL).ifPresent(control ->
+    getControl(ControlCode.SELECT_CONDITION_PANEL).ifPresent(control ->
             KeyEvents.builder(VK_S)
                     .modifiers(CTRL_DOWN_MASK)
                     .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                     .action(control)
                     .enable(this));
-    control(ControlCode.TOGGLE_CONDITION_PANEL).ifPresent(control ->
+    getControl(ControlCode.TOGGLE_CONDITION_PANEL).ifPresent(control ->
             KeyEvents.builder(VK_S)
                     .modifiers(CTRL_DOWN_MASK | ALT_DOWN_MASK)
                     .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                     .action(control)
                     .enable(this));
-    control(ControlCode.TOGGLE_FILTER_PANEL).ifPresent(control ->
+    getControl(ControlCode.TOGGLE_FILTER_PANEL).ifPresent(control ->
             KeyEvents.builder(VK_F)
                     .modifiers(CTRL_DOWN_MASK | ALT_DOWN_MASK)
                     .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                     .action(control)
                     .enable(this));
-    control(ControlCode.SELECT_FILTER_PANEL).ifPresent(control ->
+    getControl(ControlCode.SELECT_FILTER_PANEL).ifPresent(control ->
             KeyEvents.builder(VK_F)
                     .modifiers(CTRL_DOWN_MASK | SHIFT_DOWN_MASK)
                     .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                     .action(control)
                     .enable(this));
-    control(EntityTablePanel.ControlCode.PRINT).ifPresent(control ->
+    getControl(EntityTablePanel.ControlCode.PRINT).ifPresent(control ->
             KeyEvents.builder(VK_P)
                     .modifiers(CTRL_DOWN_MASK)
                     .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
@@ -792,15 +803,15 @@ public class EntityTablePanel extends JPanel {
    */
   protected final void setControl(ControlCode controlCode, Control control) {
     throwIfInitialized();
-    controls.put(requireNonNull(controlCode), control == null ? NULL_CONTROL : control);
+    controls.put(requireNonNull(controlCode), control);
   }
 
   protected Controls createToolBarControls(List<Controls> additionalToolBarControls) {
     requireNonNull(additionalToolBarControls);
     Controls toolbarControls = Controls.controls();
-    control(ControlCode.TOGGLE_SUMMARY_PANEL).ifPresent(toolbarControls::add);
-    Control toggleConditionPanelControl = control(ControlCode.TOGGLE_CONDITION_PANEL).orElse(null);
-    Control toggleFilterPanelControl = control(ControlCode.TOGGLE_FILTER_PANEL).orElse(null);
+    getControl(ControlCode.TOGGLE_SUMMARY_PANEL).ifPresent(toolbarControls::add);
+    Control toggleConditionPanelControl = getControl(ControlCode.TOGGLE_CONDITION_PANEL).orElse(null);
+    Control toggleFilterPanelControl = getControl(ControlCode.TOGGLE_FILTER_PANEL).orElse(null);
     if (toggleConditionPanelControl != null || toggleFilterPanelControl != null) {
       if (toggleConditionPanelControl != null) {
         toolbarControls.add(toggleConditionPanelControl);
@@ -810,14 +821,14 @@ public class EntityTablePanel extends JPanel {
       }
       toolbarControls.addSeparator();
     }
-    control(ControlCode.DELETE_SELECTED).ifPresent(toolbarControls::add);
-    control(ControlCode.PRINT).ifPresent(toolbarControls::add);
-    control(ControlCode.CLEAR_SELECTION).ifPresent(control -> {
+    getControl(ControlCode.DELETE_SELECTED).ifPresent(toolbarControls::add);
+    getControl(ControlCode.PRINT).ifPresent(toolbarControls::add);
+    getControl(ControlCode.CLEAR_SELECTION).ifPresent(control -> {
       toolbarControls.add(control);
       toolbarControls.addSeparator();
     });
-    control(ControlCode.MOVE_SELECTION_UP).ifPresent(toolbarControls::add);
-    control(ControlCode.MOVE_SELECTION_DOWN).ifPresent(toolbarControls::add);
+    getControl(ControlCode.MOVE_SELECTION_UP).ifPresent(toolbarControls::add);
+    getControl(ControlCode.MOVE_SELECTION_DOWN).ifPresent(toolbarControls::add);
     additionalToolBarControls.forEach(additionalControls -> {
       toolbarControls.addSeparator();
       additionalControls.actions().forEach(toolbarControls::add);
@@ -836,18 +847,18 @@ public class EntityTablePanel extends JPanel {
   protected Controls createPopupMenuControls(List<Controls> additionalPopupMenuControls) {
     requireNonNull(additionalPopupMenuControls);
     Controls popupControls = Controls.controls();
-    control(ControlCode.REFRESH).ifPresent(popupControls::add);
-    control(ControlCode.CLEAR).ifPresent(popupControls::add);
+    getControl(ControlCode.REFRESH).ifPresent(popupControls::add);
+    getControl(ControlCode.CLEAR).ifPresent(popupControls::add);
     if (popupControls.notEmpty()) {
       popupControls.addSeparator();
     }
     addAdditionalControls(popupControls, additionalPopupMenuControls);
     State separatorRequired = State.state();
-    control(ControlCode.EDIT_SELECTED).ifPresent(control -> {
+    getControl(ControlCode.EDIT_SELECTED).ifPresent(control -> {
       popupControls.add(control);
       separatorRequired.set(true);
     });
-    control(ControlCode.DELETE_SELECTED).ifPresent(control -> {
+    getControl(ControlCode.DELETE_SELECTED).ifPresent(control -> {
       popupControls.add(control);
       separatorRequired.set(true);
     });
@@ -855,7 +866,7 @@ public class EntityTablePanel extends JPanel {
       popupControls.addSeparator();
       separatorRequired.set(false);
     }
-    control(ControlCode.VIEW_DEPENDENCIES).ifPresent(control -> {
+    getControl(ControlCode.VIEW_DEPENDENCIES).ifPresent(control -> {
       popupControls.add(control);
       separatorRequired.set(true);
     });
@@ -876,7 +887,7 @@ public class EntityTablePanel extends JPanel {
       popupControls.add(columnControls);
       separatorRequired.set(true);
     }
-    control(ControlCode.SELECTION_MODE).ifPresent(control -> {
+    getControl(ControlCode.SELECTION_MODE).ifPresent(control -> {
       if (separatorRequired.get()) {
         popupControls.addSeparator();
       }
@@ -897,7 +908,7 @@ public class EntityTablePanel extends JPanel {
       addFilterControls(popupControls);
       separatorRequired.set(true);
     }
-    control(ControlCode.COPY_TABLE_DATA).ifPresent(control -> {
+    getControl(ControlCode.COPY_TABLE_DATA).ifPresent(control -> {
       if (separatorRequired.get()) {
         popupControls.addSeparator();
       }
@@ -920,7 +931,7 @@ public class EntityTablePanel extends JPanel {
             .name(Messages.print())
             .mnemonic(Messages.printMnemonic())
             .smallIcon(FrameworkIcons.instance().print());
-    control(ControlCode.PRINT).ifPresent(builder::control);
+    getControl(ControlCode.PRINT).ifPresent(builder::control);
 
     return builder.build();
   }
@@ -1196,9 +1207,9 @@ public class EntityTablePanel extends JPanel {
   private Controls createColumnControls() {
     Controls.Builder builder = Controls.builder()
             .name(MESSAGES.getString("columns"));
-    control(ControlCode.SELECT_COLUMNS).ifPresent(builder::control);
-    control(ControlCode.RESET_COLUMNS).ifPresent(builder::control);
-    control(ControlCode.COLUMN_AUTO_RESIZE_MODE).ifPresent(builder::control);
+    getControl(ControlCode.SELECT_COLUMNS).ifPresent(builder::control);
+    getControl(ControlCode.RESET_COLUMNS).ifPresent(builder::control);
+    getControl(ControlCode.COLUMN_AUTO_RESIZE_MODE).ifPresent(builder::control);
 
     return builder.build();
   }
@@ -1338,7 +1349,7 @@ public class EntityTablePanel extends JPanel {
               .action(Control.control(this::showEntityMenu))
               .enable(table);
     }
-    control(ControlCode.DELETE_SELECTED).ifPresent(control ->
+    getControl(ControlCode.DELETE_SELECTED).ifPresent(control ->
             KeyEvents.builder(VK_DELETE)
                     .action(control)
                     .enable(table));
@@ -1511,7 +1522,7 @@ public class EntityTablePanel extends JPanel {
             .name(FrameworkMessages.search())
             .smallIcon(FrameworkIcons.instance().search())
             .build();
-    control(ControlCode.CONDITION_PANEL_VISIBLE).ifPresent(conditionControls::add);
+    getControl(ControlCode.CONDITION_PANEL_VISIBLE).ifPresent(conditionControls::add);
     Controls conditionPanelControls = conditionPanel.controls();
     if (conditionPanelControls.notEmpty()) {
       conditionControls.addAll(conditionPanelControls);
@@ -1531,7 +1542,7 @@ public class EntityTablePanel extends JPanel {
             .name(FrameworkMessages.filter())
             .smallIcon(FrameworkIcons.instance().filter())
             .build();
-    control(ControlCode.FILTER_PANEL_VISIBLE).ifPresent(filterControls::add);
+    getControl(ControlCode.FILTER_PANEL_VISIBLE).ifPresent(filterControls::add);
     Controls filterPanelControls = filterPanel.controls();
     if (filterPanelControls.notEmpty()) {
       filterControls.addAll(filterPanelControls);
@@ -1602,6 +1613,10 @@ public class EntityTablePanel extends JPanel {
     }
 
     return tableConditionPanel;
+  }
+
+  private Optional<Control> getControl(ControlCode controlCode) {
+    return Optional.ofNullable(controls.get(controlCode));
   }
 
   private void configureHorizontalAlignment(ColumnConditionPanel<Attribute<?>, ?> columnConditionPanel) {